Organize longs
